requirements.txt
# based-news/requirements.txt
Flask
gunicorn
psycopg2-binary
requests
python-dotenv
nltk
google-generativeai
python-dateutil
beautifulsoup4
lxml